refactor bad statements
