+++ pyproject.toml
@@ -28,7 +28,8 @@ autoflake = "^2.3.1"
 pyright = "^1.1.399"
 jupyterlab = "^4.3.4"
 mypy = "^1.14.1"
-yamllint = "^1.37.0"
+mkdocs = "^1.6"
+mkdocs-material = "^9.5"
 
 [build-system]
 requires = ["poetry-core"]
