+++ sincpro_framework/auto_docs/__init__.py
@@ -1,10 +1,10 @@
 """
-Auto-Documentation System for Sincpro Framework
+Auto-Documentation System
 
-Sistema de documentación automática basado en introspección del framework,
-siguiendo arquitectura de domain/infrastructure con protocols de Python.
+Sistema de auto-documentación para Sincpro Framework.
 """
 
+# Domain interfaces
 from .domain import (
     ApplicationServiceMetadata,
     ComponentMetadata,
@@ -15,63 +15,12 @@ from .domain import (
     FrameworkIntrospector,
     IntrospectionResult,
 )
-from .infrastructure.service import AutoDocumentationService
 
-
-# API de conveniencia para uso externo
-def generate_framework_documentation(
-    framework_instance, output_path: str = "docs/framework_api_reference.md", **config
-) -> str:
-    """
-    Función de conveniencia para generar documentación del framework
-
-    Args:
-        framework_instance: Instancia construida del framework
-        output_path: Ruta donde guardar la documentación
-        **config: Opciones de configuración
-
-    Returns:
-        Ruta donde se guardó la documentación
-    """
-    service = AutoDocumentationService()
-    return service.generate_documentation(framework_instance, output_path, **config)
-
-
-def print_framework_summary(framework_instance) -> None:
-    """Imprime un resumen del framework en consola"""
-    service = AutoDocumentationService()
-    service.print_framework_summary(framework_instance)
-
-
-def get_framework_components(framework_instance) -> dict:
-    """Obtiene lista de componentes registrados en el framework"""
-    service = AutoDocumentationService()
-    return service.list_registered_components(framework_instance)
-
-
-def get_documentation_content(
-    framework_instance, format_type: str = "markdown", **config
-) -> str:
-    """Genera contenido de documentación sin guardar en archivo"""
-    service = AutoDocumentationService()
-    return service.generate_documentation_content(framework_instance, format_type, **config)
-
-
-__all__ = [
-    # Domain interfaces
-    "FrameworkIntrospector",
-    "DocumentationGenerator",
-    "IntrospectionResult",
-    "ComponentMetadata",
-    "DTOMetadata",
-    "FeatureMetadata",
-    "ApplicationServiceMetadata",
-    "DependencyMetadata",
-    # Infrastructure
-    "AutoDocumentationService",
-    # Public API
-    "generate_framework_documentation",
-    "print_framework_summary",
-    "get_framework_components",
-    "get_documentation_content",
-]
+# Main service
+from .service import (
+    AutoDocumentationService,
+    generate_framework_documentation,
+    get_documentation_content,
+    get_framework_components,
+    print_framework_summary,
+)
