+++ sincpro_framework/auto_docs/infrastructure/__init__.py
@@ -5,11 +5,9 @@ Implementaciones concretas de los protocols definidos en el dominio.
 """
 
 from .markdown_generator import MarkdownDocumentationGenerator
-from .service import AutoDocumentationService
 from .sincpro_introspector import SincproFrameworkIntrospector
 
 __all__ = [
-    "SincproFrameworkIntrospector",
     "MarkdownDocumentationGenerator",
-    "AutoDocumentationService",
+    "SincproFrameworkIntrospector",
 ]
